:recycle: split data_loader.py
