+++ deepchem/data/data_loader/__init__.py
@@ -1,0 +1,7 @@
+from deepchem.data.data_loader.base_loader import DataLoader
+from deepchem.data.data_loader.csv_loader import CSVLoader
+from deepchem.data.data_loader.csv_loader import UserCSVLoader
+from deepchem.data.data_loader.json_loader import JsonLoader
+from deepchem.data.data_loader.sdf_loader import SDFLoader
+from deepchem.data.data_loader.fasta_loader import FASTALoader
+from deepchem.data.data_loader.image_loader import ImageLoader
