+++ deepchem/splits/task_splitter.py
@@ -10,7 +10,6 @@ import numpy as np
 from deepchem.utils import ScaffoldGenerator
 from deepchem.utils.save import log
 from deepchem.data import NumpyDataset
-from deepchem.utils.save import load_data
 from deepchem.splits import Splitter
